+++ pkg/server.go
@@ -3,7 +3,6 @@ package garbanzo
 import (
 	"html/template"
 	"io"
-	"io/ioutil"
 	"net/http"
 
 	"github.com/labstack/echo/v4"
@@ -31,21 +30,4 @@ func NewRouter(templDir string) *echo.Echo {
 
 func rootHandler(c echo.Context) error {
 	return c.Render(http.StatusOK, "root.html", nil)
-}
-
-func NewProxyRouter() *echo.Echo {
-	e := echo.New()
-	e.GET("/*", ghHandler)
-
-	return e
-}
-
-func ghHandler(c echo.Context) error {
-	path := c.Request().URL.String()
-	url := "https://github.com" + path
-	resp, _ := http.Get(url)
-	defer resp.Body.Close()
-
-	byteArray, _ := ioutil.ReadAll(resp.Body)
-	return c.HTML(http.StatusOK, string(byteArray))
 }
